% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropls_utils.R
\name{annotate_bins}
\alias{annotate_bins}
\title{Identifica os bins referentes a cada metabolito a partir de uma planilha de anotacao de metabolitos}
\usage{
annotate_bins(integrais, loadings, use_mean = T, return_undentified = F)
}
\arguments{
\item{integrais}{Planilha de integrais, no mesmo formato utilizado para a funcao de integracao de regioes.}

\item{loadings}{planilha contendo os dados dos loadings de um objeto ropls. Gerada pela funcao extract_ropls_data(),
fornecer apenas a planilha de loadings presente no objeto do tipo list retornado.}

\item{use_mean}{Define se deve ser realizada a media dos dados para bins referntes ao mesmo metabolito.
Se TRUE (valor padrao), sera retornado apenas um bin para cada metabolito, referente ao bin medio
calculado entre todos os bins dentro da regiao de integracao do sinal, alem dos valores de vip medio
e as coordenadas do bin medio. Se FALSE, todos os bins referentes ao mesmo sinal e seus devidos dados
sao fornecidos individualmente duplicando a identificacao do metabolito quantas vezes forem necessarias.}
}
\value{
Planilha contendo apenas os bins referentes as areas dos sinais identificados e suas devidas informacoes
}
\description{
Identifica os bins referentes a cada metabolito a partir de uma planilha de anotacao de metabolitos
}
\examples{
#model_data <- extract_ropls_data(model)
#annotated_data <- annotate_bins(model_data$loadings)
}
