% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phasing.R
\name{autophase}
\alias{autophase}
\title{Aplicação Automática de Faseamento em Espectros de RMN}
\usage{
autophase(spectra, phasingMethod = "SPC_AAM", absorptionOnly = F, withBC = F)
}
\arguments{
\item{spectra}{Matriz numérica contendo os espectros de RMN a serem faseados. Cada linha representa um espectro individual.}

\item{phasingMethod}{Método de faseamento a ser utilizado. O valor padrão é \code{"SPC_AAM"}. Consulte a documentação do pacote NMRphasing para mais detalhes sobre os métodos disponíveis.}

\item{absorptionOnly}{Booleano indicando se apenas a componente de absorção deve ser considerada no faseamento. O valor padrão é \code{FALSE}.}

\item{withBC}{Booleano indicando se a correção de linha de base deve ser aplicada durante o faseamento. O valor padrão é \code{FALSE}.}
}
\value{
Retorna uma matriz contendo os espectros faseados, mantendo o mesmo formato da entrada.
}
\description{
Esta função realiza o faseamento automático de espectros de RMN utilizando o pacote NMRphasing.
}
\examples{
\dontrun{
# Supondo que "spectra_data" seja uma matriz contendo espectros de RMN
spectra_phased <- autophase(spectra_data)
}

}
