% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_PepsNMR.R
\name{calculate_spectrum_resolution}
\alias{calculate_spectrum_resolution}
\title{Calcula a resolucao dos spectros}
\usage{
calculate_spectrum_resolution(spectra, npoints = NULL, ppm = NULL)
}
\arguments{
\item{spectra}{Matriz no formato PepsNMR contendo os espectros.}

\item{npoints}{Se utilizado, a funcao passa a retornar o comprimento equivalente em deslocamento quimico
referente ao numero de pontos fornecido neste argumento. Para calcular apenas a resolucao, manter em NULL.}

\item{ppm}{Se utilizado, a funcao passa a retornar o numero de pontos no espectro correspondente ao
comprimento em ppm forncecido. Para calcular apenas a resolucao, manter em NULL.}
}
\value{
Um valor ou string informando a resolucao ou as conversoes entre ppm para numero de pontos e
numero de pontos para ppm.
}
\description{
Funcao auxiliar criada para calcular a resolucao da matriz de espectros fonecida.
}
\examples{
#calculate_spectrum_resolution(spec)
#calculate_spectrum_resolution(spec,npoints = 128)
#calculate_spectrum_resolution(spec,ppm=0.04)

}
