% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanData.R
\name{CleanDataMatrix}
\alias{CleanDataMatrix}
\title{Executa a remocao de valores constantes e invalidos (NA) nas matrizes de preprocessamento ou dados para estatisitica}
\usage{
CleanDataMatrix(ndata)
}
\arguments{
\item{ndata}{Recebe uma matriz do tipo numerica a ser "limpa". A planilha deve conter apenas as variaveis a serem
filtradas, nao contendo nenhum metadado ou variavel do tipo categorica/Factor/Character.}
}
\value{
Objeto do tipo Matrix, contendo o dado filtrado sem as colulas constantes/invalidas.
}
\description{
Executa a remocao de valores constantes e invalidos (NA) nas matrizes de preprocessamento ou dados para estatisitica
}
\examples{
\dontrun{
planilha_filtrada <- CleanDataMatriz(planilha)
}
}
