% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phasing.R
\name{compare_phasing_methods}
\alias{compare_phasing_methods}
\title{Comparação de Métodos de Faseamento para Espectros de RMN}
\usage{
compare_phasing_methods(
  spectra,
  fid_info,
  ref = 1,
  absorptionOnly = F,
  withBC = F,
  phasingMethods = c("NLS", "MPC_DAOM", "MPC_EMP", "MPC_AAM", "MPC_DSM", "MPC_ADSM",
    "SPC_DAOM", "SPC_EMP", "SPC_AAM", "SPC_DSM", "SPC_ADSM")
)
}
\arguments{
\item{spectra}{Matriz numérica contendo os espectros de RMN a serem analisados. Cada linha representa um espectro individual.}

\item{fid_info}{Informações do domínio do tempo (FID) necessárias para o referenciamento interno dos espectros faseados.}

\item{ref}{Índice da linha na matriz \code{spectra} que será utilizada como espectro de referência para a comparação dos métodos de faseamento. O valor padrão é \code{1}.}

\item{absorptionOnly}{Booleano indicando se apenas a componente de absorção deve ser considerada no faseamento. O valor padrão é \code{FALSE}.}

\item{withBC}{Booleano indicando se a correção de linha de base deve ser aplicada durante o faseamento. O valor padrão é \code{FALSE}.}

\item{phasingMethods}{Vetor de caracteres contendo os nomes dos métodos de faseamento a serem testados. Os métodos disponíveis incluem \code{"NLS"}, \code{"MPC_DAOM"}, \code{"MPC_EMP"}, \code{"MPC_AAM"}, \code{"MPC_DSM"}, \code{"MPC_ADSM"}, \code{"SPC_DAOM"}, \code{"SPC_EMP"}, \code{"SPC_AAM"}, \code{"SPC_DSM"}, e \code{"SPC_ADSM"}. O valor padrão inclui todos esses métodos.}
}
\value{
Retorna uma matriz contendo os espectros de referência faseados por cada um dos métodos especificados, juntamente com o espectro faseado utilizando \code{PepsNMR::ZeroOrderPhaseCorrection}.
}
\description{
Esta função aplica e compara diferentes métodos de faseamento a um espectro de RMN de referência.
}
\examples{
\dontrun{
# Supondo que "spectra_data" seja uma matriz contendo espectros de RMN e "fid_info_data" as informações do FID
phased_spectra <- compare_phasing_methods(spectra_data, fid_info_data)
}

}
