% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropls_utils.R
\name{extract_ropls_data}
\alias{extract_ropls_data}
\title{Extrai no formato de planilha as informacoes de um modelo do pacote ropls}
\usage{
extract_ropls_data(ropls_object)
}
\arguments{
\item{ropls_object}{Objeto da classe opls gerado pelo pacote ropls. Pode ser de qualquer um dos tipos: PCA, PLS-DA ou OPLS-DA.}
}
\value{
Retorna uma lista contendo dois dataframes. Um referente aos scores do modelo gerado e outro
referente aos loadings
}
\description{
Extrai em um objeto os scores e loadings de um modelo de PCA, PLS-DA ou OPLS-DA gerados pelo
pacote ropls. A deteccao do tipo de modelo e executada automaticamente necessitando apenas fornecer
o objeto gerado pelo pacote ropls.
}
\examples{
#model <- opls(xdada,yvector, predI = 1, orthoI = 1)
#model_data <- extract_ropls_data(model)
}
