% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_annotation.R
\name{filter_annotated_bins}
\alias{filter_annotated_bins}
\title{Filtra e anota bins espectrais}
\usage{
filter_annotated_bins(
  bins_data,
  annotation_data,
  ignore_columns,
  number_of_decimal_places = 4,
  long_format = T
)
}
\arguments{
\item{bins_data}{Data frame contendo os dados espectrais, onde as colunas (exceto as especificadas em \code{ignore_columns}) representam os bins espectrais.}

\item{annotation_data}{Lista ou data frame contendo duas colunas: a primeira com os nomes dos metabólitos e a segunda com os bins correspondentes.}

\item{ignore_columns}{Vetor de caracteres especificando as colunas do \code{bins_data} que não devem ser consideradas como bins espectrais.}

\item{number_of_decimal_places}{Inteiro indicando o número de casas decimais a serem consideradas para arredondamento dos bins. Padrão é 4.}

\item{long_format}{Lógico. Se \code{TRUE}, retorna os dados em formato longo; se \code{FALSE}, retorna em formato largo. Padrão é \code{TRUE}.}
}
\value{
Um data frame contendo os dados filtrados e anotados. O formato do retorno depende do argumento \code{long_format}.
}
\description{
Esta função filtra um conjunto de dados espectrais com base em anotações fornecidas e pode retornar os dados em formato longo ou largo.
}
\examples{
# Exemplo de uso:
bins_data <- data.frame(SampleID = c("S1", "S2"), "1.2345" = c(0.5, 0.6), "2.3456" = c(0.3, 0.4),check.names = FALSE)
annotation_data <- data.frame(Metabolito = c("Met1", "Met2"), bins = c(1.2345, 2.3456))
filter_annotated_bins(bins_data, annotation_data, ignore_columns = "SampleID")

}
