% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Integration.R
\name{hibrid_bucketing}
\alias{hibrid_bucketing}
\title{Permite gerar uma matriz hibrida entre bins equidistantes e regioes de integracao personalizadas}
\usage{
hibrid_bucketing(spectra, integrals, bin_width = 0.01)
}
\arguments{
\item{spectra}{Matrix contendo os espectros processados de acordo com o pacote PepsNMR.}

\item{integrals}{Dataframe derivado da planilha contendo as regioes de integracao. Utilizar o mesmo modelo exigido para a funcao
NMR_integration()}

\item{bin_width}{Largura de bin aplicada para as regioes nao inclusas nas regioes determinadas no dataframe "integrals"}
}
\value{
Um dataframe contendo as integrais e bins gerados.
}
\description{
Esta funcao gera um hibrido entre um sistema de integrais definidas pelo usuario e bins classicos de larguras equidistantes
Para todas as regioes que estiverm compreendidas no intevalo das integrais fonecidas, estas serao identificadas de acordo com
o metabolito sinalizado no dataframe integrals, enquanto todas as demais regioes serao processadas com bins de largura definida
no argumento bin_width. Para facilitar a leitura, todas as integrais sao movidas para o inicio da matriz, alem disso, as regioes
integradas sao removidas dos bins equidistantes para evitar duplicatas de variaveis.
}
\examples{
#NMR_hbins <- hibrid_bucketing(spectra = spec,integrals = regioes_integrais,bin_width = 0.01)
}
