% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Integration.R
\name{NMR_integration}
\alias{NMR_integration}
\title{Calcula as integrais a partir de uma planilha processada de espectros de RMN, recomendada para uso com dados processados pelo pacote PepsNMR}
\usage{
NMR_integration(spectral_matrix, integration_intervals)
}
\arguments{
\item{spectral_matrix}{Dataframe gerado do processamento dos dados espectrais, contendo a identificação das amostras na primeira coluna e
as intensidades dos deslocamentos químicos nas demais.}

\item{integration_intervals}{Dataframe formatado a partir do Excel (preferencialmente), contendo a identificação dos metabolitos, inicio do intervalo para integração
e fim do intervalo para integração, respectivamente.}
}
\value{
Um data frame contendo o valor da area das integrais no formato amostras nas linhas e metabolitos nas colunas
}
\description{
Integra as regiões fornecidas no argumento integration_intervals, preferencialmente deve ser um dataframe produzido a partir de uma planilha do excel.
Deve obrigatoriamente conter 3 colunas, sendo a primeira a identificação dos metabolitos, a segunda o deslocamento quimico inicial de cada integral e a terceira o
deslocamento quimico final de cada integral. Os nomes de cada coluna pode ser escolhido livremente, mas a ordem deve permancer a mesma.
}
\details{
O dataframe utilizado como input deve ser gerado a partir da funcao pepsMatrixToDF() ou possuir uma estrutura compativel.
}
\examples{
#Exemplo de dataframe no formato correto para o argumento "Integration_intervals"
 region <- data.frame(metabolito=c("Leucina","Etanol","VLDL","metilhistidina","teste"),ppm_i=c(0.94,1.1,0.81,3.685,3.30),ppm_f=c(0.97,1.3,0.94,3.7,3.31))
}
