% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_PepsNMR.R
\name{pepsMatrixToDF}
\alias{pepsMatrixToDF}
\title{Converte o objeto matrix gerado pelo PepsNMR em um dataframe formatado}
\usage{
pepsMatrixToDF(Pepsmatrix, DirNames = F)
}
\arguments{
\item{Pepsmatrix}{Matrix gerada por alguma das etapas de processamento do PepsNMR}

\item{DirNames}{Boleano, indicando se o nome das amostras equivale ao numero do experimento. Se verdadeiro (TRUE) realiza a correção
da ordem alfanumerica dos numeros das amostras.}
}
\value{
Returns a formated dataframe
}
\description{
Esta função recebe o objeto retornado pelas funções de pre-processamento do pacote PepsNMR e retorna um objeto do tipo dataframe
formatado de maneira adequada para salvar a planilha ou ser utilizado em outras funções. Além disso, também faz a correção da ordem
alfabetica nos nomes das amostras.
}
\examples{
\dontrun{NMR_dataproc <- pepsMatrixToDF(Spectrum_data.N)}
}
