% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_PepsNMR.R
\name{plot_interactive_spectra}
\alias{plot_interactive_spectra}
\title{Plota um grafico interativo a partir de qualquer um dos objetos gerados durante o processamento via PepsNMR}
\usage{
plot_interactive_spectra(
  Spectrum_data,
  plot_resolution = 0.25,
  limit_n_points = F,
  plot_only = 0,
  Spectrum_window = NULL
)
}
\arguments{
\item{Spectrum_data}{Objeto contendo os dados espectrais gerados a partir do PepsNMR. Pode estar na forma de matrix ou dataframe.}

\item{plot_resolution}{Caso o argumento limit_n_points seja TRUE, define para quantos porcento do número total de de pontos a resolução dos espectros será reduzida.
Recomendado usar valores multiplos de 0.25 para evitar serrilhamento do espectro.}

\item{limit_n_points}{Booleano, definindo se os espectros devem ou não sofre redução da resolução/numero de pontos antes de serem plotados.
Recomendado na visualização de multiplos espectros sem redução da janela espectral.}

\item{plot_only}{Vetor numerico definindo os indices dos espectros a serem plotados (Pode ser um unico indice ou vetor contendo multiplos indices)
. Se definido como zero, apenas os 10 primeiros
espectros serão representados (Padrão) ou o numero total de espectros se menor que 10.}

\item{Spectrum_window}{Recebe um vetor de tamanho 2, contendo os limites superior e inferior da janela espectral a ser representada}
}
\value{
Retorna um objeto de grafico interativo do tipo plotly
}
\description{
Plota um grafico interativo a partir de qualquer um dos objetos gerados durante o processamento via PepsNMR
}
\examples{
\dontrun{Plot_NMR <- plot_interactive_Spectra(Bins_data)}
}
