% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROPLS_plots.R
\name{Plot_loading}
\alias{Plot_loading}
\title{Plot Loadings for Multivariate Models}
\usage{
Plot_loading(
  model,
  comp = c(1, 2),
  point = T,
  repel = F,
  font.label = c(12, "plain"),
  bin_identification = NULL,
  bins_roundPrecision = 4,
  VIP_roundPrecision = 2,
  show_metabolite_name = T,
  VIP_values_scale = NULL,
  theme = NULL
)
}
\arguments{
\item{model}{An object of class \code{ropls}, representing the multivariate model (PCA, PLS(-DA), or OPLS(-DA)) for which loadings are plotted.}

\item{comp}{A numeric vector of length 2 indicating the components to be plotted (default: \code{c(1, 2)}).}

\item{point}{Logical. If \code{TRUE}, points are displayed in the plot (default: \code{TRUE}).}

\item{repel}{Logical. If \code{TRUE}, labels are adjusted to repel each other, reducing overlap (default: \code{FALSE}).}

\item{font.label}{A vector indicating the font size and style of labels (e.g., \code{c(12, "plain")}).}

\item{bin_identification}{A data frame with two columns: the first column for metabolite names and the second for corresponding bins. Used to annotate bins.}

\item{bins_roundPrecision}{Integer. Specifies the number of decimal places for rounding bin values (default: \code{4}).}

\item{VIP_roundPrecision}{Integer. Specifies the number of decimal places for rounding VIP (Variable Importance in Projection) values (default: \code{2}).}

\item{show_metabolite_name}{Logical. If \code{TRUE}, metabolite names are displayed in labels when \code{bin_identification} is provided (default: \code{TRUE}).}

\item{VIP_values_scale}{A numeric vector of values to scale the color gradient for VIP values. If \code{NULL}, a default scale is applied. Use a vector of length 3 were the first value represents
the lower point of the scale, followed by the middle point and the highest point.}

\item{theme}{A ggplot2 theme to customize the plot appearance. If \code{NULL}, the default \code{theme_bw} is used.}
}
\value{
A ggplot object representing the loadings plot for the specified multivariate model.
}
\description{
This function generates a scatter plot of loadings for multivariate models, including PCA, PLS(-DA), and OPLS(-DA).
The function allows for customization of various plot parameters, enabling the identification and visualization
of important features (e.g., bins or metabolites) within the model. It supports options to label identified bins,
control the precision of rounding for bin and VIP values, and apply custom themes.
}
\examples{
\dontrun{
# Example usage:
model <- some_ropls_model   # Replace with your model object
bin_info <- data.frame(Metabolite = c("Met1", "Met2"), Bins = c(0.12, 0.34))
Plot_loading(model, comp = c(1, 2), point = TRUE, bin_identification = bin_info)
}

}
