% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROPLS_plots.R
\name{Plot_loading2}
\alias{Plot_loading2}
\title{Plot Loadings of a Multivariate Model}
\usage{
Plot_loading2(
  model,
  comp = c(1, 2),
  point = T,
  repel = F,
  font.label = c(12, "plain"),
  bins_roundPrecision = 4,
  VIP_roundPrecision = 2,
  VIP_values_scale = NULL,
  theme = NULL
)
}
\arguments{
\item{model}{An object containing a PCA, PLS, PLS-DA, OPLS, or OPLS-DA model from the ROPLS package.}

\item{comp}{A numeric vector of length 2 specifying the components to be plotted. Default is c(1,2).}

\item{point}{Logical. If TRUE, points representing variables are displayed. Default is TRUE.}

\item{repel}{Logical. If TRUE, labels are repelled to avoid overlap. Default is FALSE.}

\item{font.label}{A vector specifying font size, style, and color for variable labels. See \code{ggscatter} from \code{ggpubr} for details.}

\item{bins_roundPrecision}{Numeric. Number of decimal places for rounding bin values. Default is 4.}

\item{VIP_roundPrecision}{Numeric. Number of decimal places for rounding VIP values. Default is 2.}

\item{VIP_values_scale}{Numeric vector. Scale values for VIP coloring gradient.Use a vector of length 3 were the first value represents
the lower point of the scale, followed by the middle point and the highest point.}

\item{theme}{A \code{ggplot2} theme object for customizing the plot appearance.}
}
\value{
A \code{ggplot2} object displaying the loading plot.
}
\description{
This function generates a loading plot for PCA, PLS, PLS-DA, OPLS, or OPLS-DA models.
It visualizes the importance of variables in different components and should be used
instead of \code{Plot_scores} when the predictor variables in the model do not correspond to a spectrum.
}
\details{
\itemize{
\item For PCA models, variables are colored uniformly.
\item For PLS and PLS-DA models, variables are colored according to their VIP scores.
\item For OPLS and OPLS-DA models, predictive (\code{p}) and orthogonal (\code{o}) components are correctly labeled on the axes.
}

This function should be used instead of \code{Plot_scores} when the variables used as predictors in the model
do not correspond to a spectrum.
}
\examples{
\dontrun{
Plot_loading2(model, comp = c(1,2), point = TRUE, repel = FALSE, font.label = c(12, "plain"))
}
}
