% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raster.R
\name{plot_raster_spectrum}
\alias{plot_raster_spectrum}
\title{Plota os espectros na vertical}
\usage{
plot_raster_spectrum(
  spec_data,
  plot_region = c(0, 9),
  spec_resolution = 0.5,
  plot_resolution = 0.02
)
}
\arguments{
\item{spec_data}{Matriz gerada pelo PepsNMR contendo os espectros de RMN processados}

\item{plot_region}{Recebe um vetor onde o primeiro valor se refere ao inicio do intervalo e o segundo ao final do intervalo a ser plotado.
Delimita a subregião do espectro a ser plotada, util para reduzir o tamanho do grafico.}

\item{spec_resolution}{Define a resolucao a ser conservada nos espectros. Recebe um valor entre 1 e 0, onde 1 representa a resolucao original
e qualquer valor entre 1 e 0 o percentual da resolucao a ser mantido. Ex: Uma resolucao de 0.5 (Valor padrao) ira manter 16000 pontos em um espectro contendo
32000 pontos.}

\item{plot_resolution}{Define a resolucao do gradiente de cor. recebe valores entre 1 e 0, onde valores maiores favorecem a representação de sinais mais intensos
e promove maior supressao do ruido na representacao e valores menores favorece sinais menos intensos, mas aumenta o ruido observado. Padronizado em 0.02}
}
\value{
Retorna um grafico interativo no formato de objeto da classe plotly.
}
\description{
Plota os espectros projetados com vista superior. Util para localizar sinais isolados/discrepantes em apenas uma amostras ou delimitar a
area de corte para as integrais. No momento, exige a opcao dir.names = T, na funcao ReadFids do pacote PepsNMR.
}
\examples{
\dontrun{
plot_raster_spectrum(spec_data = Spectrum_data.WS[-3,],plot_region = c(0.5,4),plot_resolution = 0.03,spec_resolution = 0.1)
}

}
