% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROPLS_plots.R
\name{Plot_scores}
\alias{Plot_scores}
\title{Plot Scores for Multivariate Models}
\usage{
Plot_scores(
  model,
  groups = NULL,
  comp = c(1, 2),
  point_size = 2,
  ellipse = T,
  labels = T,
  font.label = c(12, "plain"),
  point = T,
  repel_labels = F,
  theme = NULL
)
}
\arguments{
\item{model}{A list containing a model of type PCA, PLS-DA, OPLS-DA, PLS, or OPLS, generated by the \code{ropls} package.}

\item{groups}{A vector specifying sample groups used for coloring and identifying samples in PCA models.
In supervised models (PLS-DA, OPLS-DA), the groups are automatically defined by the model.}

\item{comp}{A numeric vector of length 2 specifying the components to be plotted. Defaults to \code{c(1, 2)}.}

\item{point_size}{Numeric. Defines the size of the points in the plot (default: \code{2}).}

\item{ellipse}{Logical. If \code{TRUE}, confidence interval ellipses are displayed around groups (default: \code{TRUE}).}

\item{labels}{Logical or vector. If \code{TRUE}, sample names are displayed as labels.
Alternatively, a custom vector of sample names can be provided. If \code{FALSE}, no labels are shown (default: \code{TRUE}).}

\item{font.label}{A vector specifying the font size, color, and style for sample labels. For details, see
the \code{font.label} argument in the \code{ggscatter} function from the \code{ggpubr} package.}

\item{point}{Logical. If \code{TRUE}, points representing samples are displayed in the plot (default: \code{TRUE}).}

\item{repel_labels}{Logical. If \code{TRUE}, sample labels are adjusted to avoid overlap (default: \code{FALSE}).}

\item{theme}{A ggplot2 theme for customizing the plot appearance. If \code{NULL}, the default theme \code{theme_bw} is used.}
}
\value{
A \code{ggplot2} object representing the scores plot for the specified multivariate model.
}
\description{
Provides an easy way to generate scores plots from models generated by the \code{ropls} package,
with advanced customization options such as group-based coloring, confidence ellipses,
and flexible label positioning.
}
\details{
This function generates a scores plot for multivariate models created using the \code{ropls} package,
including PCA, PLS(-DA), and OPLS(-DA). It provides advanced customization options for visualization
using \code{ggplot2}, allowing the addition of ellipses, sample labels, and group-based coloring.
}
\examples{
\dontrun{
# Example usage:
model <- some_ropls_model   # Replace with your PCA, PLS-DA, or OPLS-DA model
groups <- c("Group1", "Group2", "Group1", "Group2")  # Define sample groups for coloring
Plot_scores(model, groups = groups, comp = c(1, 2), ellipse = TRUE, labels = TRUE)
}

}
