% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROPLS_plots.R
\name{plot_scores_overview}
\alias{plot_scores_overview}
\title{Generates multiple PCA score plots}
\usage{
plot_scores_overview(
  PCA_object,
  groups,
  labels,
  label_size = 7,
  plot_font_size = 7,
  repel = F,
  show_points = T,
  add = NULL,
  scale_expansion = 0.15,
  return_list = F
)
}
\arguments{
\item{PCA_object}{ropls PCA object containing the analysis results.}

\item{groups}{Vector of groups to color the points in the plots.}

\item{labels}{Vector of labels to identify the points in the plots.}

\item{label_size}{Font size for the labels. Default is 7.}

\item{plot_font_size}{Font size for the plot elements. Default is 7.}

\item{repel}{Logical. If \code{TRUE}, labels are repelled to avoid overlap. Default is \code{FALSE}.}

\item{show_points}{Logical. If \code{TRUE}, points are displayed in the plot. Default is \code{TRUE}.}

\item{add}{Additional elements to be added to the plots, such as scales, labels, or any other
ggplot2 element. For multiple elements provide a list containing the elements to be added. Default is \code{NULL}.}

\item{scale_expansion}{Axis expansion factor for the plots, used to avoid cliping sample names on plot margins. Default is 0.15.}

\item{return_list}{Logical. If \code{TRUE}, returns a list containing the ggplot2 objects to be used with wrap_plots().
if \code{FALSE} return a single combined plot. Default is \code{FALSE}.}
}
\value{
A single plot combining multiple pairs of principal components or a list of plots if \code{return_list = TRUE}.
}
\description{
This function creates an overview of PCA scores, generating plots for each pair of principal components.
}
\examples{
\dontrun{
 Example usage:
 plot_scores_overview(PCA_object, groups = my_groups, labels = my_labels)
}

}
