% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROPLS_plots.R
\name{plot_slines}
\alias{plot_slines}
\title{Plot S-Lines for PLS(-DA) Models}
\usage{
plot_slines(TrainingData, model, comp = 1, interactive = T)
}
\arguments{
\item{TrainingData}{A matrix or data frame containing the original data used for model training.
Variables are represented as columns, and samples as rows.}

\item{model}{A OPLS(-DA), PLS(-DA) or PCA model generated using the \code{ropls} package.}

\item{comp}{A numeric value specifying the component to be analyzed. Default is \code{1}.}

\item{interactive}{Logical. If \code{TRUE}, returns an interactive plot using \code{plotly}.
If \code{FALSE}, returns a static \code{ggplot2} plot. Default is \code{TRUE}.}
}
\value{
A plot object. Returns a \code{ggplot2} object for static plots or a \code{plotly} object for interactive plots.
}
\description{
Creates an S-line plot to show the relationship between variables and the scores of a selected
component. The plot displays the covariance (\code{Cov(Tp, X)}) on the y-axis and the chemical shift (ppm)
on the x-axis. The fill color represents the absolute correlation (\verb{|Cor(Tp, X)|}),
highlighting variables that contribute most to the selected component.
}
\details{
Generates S-line plots to visualize the covariance and correlation of variables
with the scores of a selected component from a OPLS(-DA), PLS(-DA) or PCA model.
This plot provides insight into the contribution of each variable to the selected component.
}
\examples{
\dontrun{
# Example usage:
TrainingData <- some_data_matrix  # Replace with your dataset
model <- some_ropls_model         # Replace with your PLS-DA or PCA model
plot_slines_PLS(TrainingData, model, comp = 1, interactive = TRUE)
}

}
