% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_PepsNMR.R
\name{Spectra_align}
\alias{Spectra_align}
\title{Alinhamento espectral via metodo CluPA}
\usage{
Spectra_align(
  spectra,
  nDivRange_ppm = 0.04,
  maxShifts_ppm = 0.015,
  baselineThreshold = 5e+06,
  SNR.Th = -1,
  show_info = T
)
}
\arguments{
\item{spectra}{Matriz de espectros gerada pelo PepsNMR, contendo os deslocamentos quimicos nas colunas e os espectros nas linhas.
OBS: Não realizar a conversao usando a funcao "PepsMatrixtoDF()"}

\item{nDivRange_ppm}{Tamanho de bin (em ppm) utilizado para a segmentacao do espectro durante o processo de identificacao dos picos.}

\item{maxShifts_ppm}{Deslocamento maximo permitido (em ppm) durante o alinhamento para cada sinal identificado. Pode ser utilizado o valor "auto"
para que o algoritmo determine automaticamente o deslocamento ideal.}

\item{baselineThreshold}{Linha de corte de intesidade para a deteccao dos picos. Determinar com base na inspecao dos espectros.}

\item{SNR.Th}{Razao sinal/Ruido considerada durante a deteccao dos picos. Pode receber o valor -1 onde o algoritmo determina automicaticamente a
razao sinal/ruido}

\item{show_info}{Recebe True/False. Configura se as mensagens de progresso devem ser mostradas durante o processo.}
}
\value{
Uma matriz no mesmo formato do dado de entrada (Porem contendo apenas a parte real do espectro) ja alinhada.
}
\description{
Aplica o metodo hierarchical Cluster-based Peak Alignment (CluPA) aos espectros, seguindo os parametros fonecidos pelo usuario.
Esta funcao foi otimizada para funcionar a partir dos espectros processados com o pacote PepsNMR, podendo ter comportamentos
diferentes do previsto se utilizada com dados de outras ferramentas.
Para detalhes extras, ver a documentacao do pacote "speaq" onde o algoritmo usado aqui foi implementado originalmente.
}
\examples{
#aling_spec1 <- Spectra_align(spec)
}
