% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VennDiagram.R
\name{Venn_diagram}
\alias{Venn_diagram}
\title{Title Cria um diagrama de Venn a partir dos de uma lista fornecida}
\usage{
Venn_diagram(
  x,
  fill = RColorBrewer::brewer.pal(length(x), "Dark2"),
  color = RColorBrewer::brewer.pal(length(x), "Dark2"),
  alp = 0.5,
  heig = 9000,
  widt = 9000,
  title = NULL,
  cat_names = names(x),
  cat_font_size = 1,
  var_font_size = 1
)
}
\arguments{
\item{x}{Lista contendo o conteudo ao qual o Diagrama deve ser produzido. Cada objeto da lista deve ser um vetor de strings com os membros do grupo.}

\item{fill}{Vetor de cores a ser usado no preenchimento do diagrama. Deve ter o mesmo tamanho da lista com os grupos.}

\item{color}{Vetor de cores a ser usado no contornos do diagrama. Deve ter o mesmo tamanho da lista com os grupos.}

\item{alp}{Transparencia do grafico}

\item{heig}{Altura da figura gerada (Em pixels)}

\item{widt}{Largura da figura gerada (Em pixels)}

\item{title}{Titulo do grafico}

\item{cat_names}{Vetor de nomes a serem atribuidos aos grupos no diagrama}

\item{cat_font_size}{Tamanho da fonte do nome dos grupos}

\item{var_font_size}{Tamanho da fonte do nome dos membros dos grupos}
}
\value{
Um plot contendo o diagrama de Venn na forma de objeto de figura do R
}
\description{
Cria um diagrama de Venn a partir dos de uma lista fornecida com um objeto (Vetor) para cada grupo
}
\examples{
 x <- list()
 x$Male <- c("1-Methyl-Histidine","L-Lactic acid","D-Glucose","VLDL & LDL")
 x$Female <- c("1-Methyl-Histidine","L-Glutamic acid","Acetone","D-Glucose")
 x$Adults <- c("L-Lactic acid","VLDL & LDL","1-Methyl-Histidine","Acetone","Acetoacetic acid")
 x$Elderly <- c("D-Glucose","1-Methyl-Histidine","L-Glutamic acid","Acetone")

 graph1 <- Venn_diagram(x)

}
