% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_PepsNMR.R
\name{plot_interactive_Spectra}
\alias{plot_interactive_Spectra}
\title{Plota um grafico interativo a partir de qualquer um dos objetos gerados durante o processamento via PepsNMR}
\usage{
plot_interactive_Spectra(
  Spectrum_data,
  plot_resolution = 0.25,
  limit_n_points = T,
  plot_only = 0
)
}
\arguments{
\item{Spectrum_data}{Objeto contendo os dados espectrais gerados a partir do PepsNMR (Normalmente será uma matrix)}

\item{plot_resolution}{Caso o argumento limit_n_points seja TRUE, define para quantos porcento do número total de de pontos a resolução dos espectros será reduzida.
Recomendado usar valores multiplos de 0.25 para evitar serrilhamento do espectro.}

\item{limit_n_points}{Boleano, definindo se os espectros devem ou não sofre redução da resolução/numero de pontos antes de serem plotados.
Recomendado caso possua muitos espectros ou esteja rodando em uma maquina mais fraca. Por padrão, a resolução será limitada a 10000 pontos/espectro}

\item{plot_only}{Vetor numerico definindo os indices dos espectros a serem plotados. Se definido como zero, todos os espectros na matrix serao plotados (Padrao).
Por conta da coersao automatica do tipo de dado, nao e possivel plotar apenas um espectro, de forma que se for fornecido apenas um valor o espectro seguinte sera plotado automaticamente.}
}
\value{
Retorna um objeto de grafico interativo do tipo plotly
}
\description{
Plota um grafico interativo a partir de qualquer um dos objetos gerados durante o processamento via PepsNMR
}
\examples{
\dontrun{Plot_NMR <- plot_interactive_Spectra(Bins_data)}
}
