% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROPLS_plots.R
\name{Plot_scores}
\alias{Plot_scores}
\title{Title}
\usage{
Plot_scores(
  model,
  groups = NULL,
  comp = c(1, 2),
  point_size = 2,
  ellipse = T,
  labels = T,
  font.label = c(12, "plain"),
  point = T,
  repel_labels = F,
  theme = ggpubr::theme_pubr()
)
}
\arguments{
\item{model}{Lista contendo um modelo de PCA, PLS-DA, OPLS-DA, PLS, ou OPLS gerado a partir
do pacote ROPLS.}

\item{groups}{Utilizado em modelos de PCA, recebe um vetor de grupos utilizado para colorir e
identificar as amostras em diferentes grupos. Nos modelos supervisionados os grupos são definidos
pelo proprio modelo.}

\item{comp}{Vetor de tamanho 2 especificando as componentes a serem representadas no grafico. Por
padrão as componentes 1 e 2 serão mostradas.}

\item{point_size}{Tamanho dos pontos na representação. (Booleano)}

\item{ellipse}{Booleano utilizado para definir se a elipse do intervalo de confiânça deverá
ser mostrada ou não.}

\item{labels}{booleano controlando se o nome das amostras deve ser presentado ou não.Também pode
receber diretamente um vetor de nomes para as amostras, substituindo os nomes das amostras fornecidos pelo modelo}

\item{font.label}{Recebe um vetor contendo o tamanho de fonte, cor de fonte e tipo de fonte usada
nos nomes das amostras. Para mais detalhes ver documentação do argumento de mesmo nome na função
ggscatter do pacote ggpubr.}

\item{point}{Booleano definindo se devem ser representados pontos para cada amostra.}

\item{repel_labels}{Utilizado caso as amostras possuam rotulos em sobreposição, tenta reorganizar as
posições dos rotulos para permitir a visualização de todos.}

\item{theme}{Controla o tema do grafico, pode ser utilizado qualquer um dos temas do ggplot2.}
}
\value{
Um grafico do ggplot2 contendo os scores do modelo em questão.
}
\description{
Fornece opções para plotar os graficos de scores dos modelos gerados pelo pacote ROPLS com opções
mais avançadas de personalização baseadas no ggplot2.
}
